+++ revenuecat_examples/purchase_tester/.run/Amazon Appstore.run.xml
@@ -1,6 +1,7 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="Amazon Appstore" type="FlutterRunConfigurationType" factoryName="Flutter">
-    <option name="filePath" value="$PROJECT_DIR$/lib/main_amazon_appstore.dart" />
+    <option name="additionalArgs" value="--dart-define=AMAZON=true" />
+    <option name="filePath" value="$PROJECT_DIR$/lib/main.dart" />
     <method v="2" />
   </configuration>
 </component>
