+++ revenuecat_examples/purchase_tester/lib/main.dart
@@ -1,0 +1,24 @@
+import 'dart:io';
+
+import 'package:flutter/material.dart';
+import 'package:purchases_flutter_example/src/constant.dart';
+
+import 'store_config.dart';
+import 'src/app.dart';
+
+void main() {
+  if (Platform.isIOS || Platform.isMacOS) {
+    StoreConfig(
+      store: Store.appleStore,
+      apiKey: appleApiKey,
+    );
+  } else if (Platform.isAndroid) {
+    const useAmazon = bool.fromEnvironment("amazon");
+    StoreConfig(
+      store: useAmazon ? Store.amazonAppstore : Store.googlePlay,
+      apiKey: useAmazon ? amazonApiKey : googleApiKey,
+    );
+  }
+
+  runApp(const PurchaseTester());
+}
