--- revenuecat_examples/purchase_tester/lib/main_apple_store.dart
@@ -1,14 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:purchases_flutter_example/src/constant.dart';
-
-import 'store_config.dart';
-import 'src/app.dart';
-
-void main() {
-  StoreConfig(
-    store: Store.appleStore,
-    apiKey: appleApiKey,
-  );
-
-  runApp(const PurchaseTester());
-}
